Add the installation directory of pithos on Debian to the PYTHONPATH
Debian installs pithos to /usr/share/pithos instead of
/usr/lib/python3/dist-packages. As I don't want to replicate its code
in rhythmbox-pandora (and having to keep track of changes) but want to
use the system-installed version, I have to add the directory to PYTHONPATH
in order to access its modules.

On Fedora/RedHat/OpenSuSE, the installation path seems to match a
PYTHONPATH, so this is not necessary.